begin work on route stack component
